Clean up initialization on tests.ts
