+++ src/tests/clipperUI/components/modeButton_tests.tsx
@@ -1,6 +1,5 @@
 import {ModeButton} from "../../../scripts/clipperUI/components/modeButton";
 
-import {HelperFunctions} from "../../helperFunctions";
 import {MithrilUtils} from "../../mithrilUtils";
 import {MockProps} from "../../mockProps";
 import {TestModule} from "../../testModule";
