+++ src/tests/clipperUI/components/previewViewer/previewViewerPdfHeader_tests.tsx
@@ -5,7 +5,6 @@ import {PreviewViewerPdfHeader, PreviewViewerPdfHeaderProp} from "../../../../sc
 
 import {Constants} from "../../../../scripts/constants";
 
-import {HelperFunctions} from "../../../helperFunctions";
 import {MithrilUtils} from "../../../mithrilUtils";
 import {MockProps} from "../../../mockProps";
 import {TestModule} from "../../../testModule";
