+++ src/tests/clipperUI/panels/changeLogPanel_tests.tsx
@@ -3,7 +3,6 @@ import {Constants} from "../../../scripts/constants";
 import {TooltipProps} from "../../../scripts/clipperUI/tooltipProps";
 import {ChangeLogPanel} from "../../../scripts/clipperUI/panels/changeLogPanel";
 
-import {HelperFunctions} from "../../helperFunctions";
 import {MithrilUtils} from "../../mithrilUtils";
 import {MockUpdates} from "../../mockUpdates";
 import {TestModule} from "../../testModule";
