+++ src/tests/clipperUI/panels/ratingsPanel_tests.tsx
@@ -14,7 +14,6 @@ import {Constants} from "../../../scripts/constants";
 import {ObjectUtils} from "../../../scripts/objectUtils";
 import {Settings} from "../../../scripts/settings";
 
-import {HelperFunctions} from "../../helperFunctions";
 import {MithrilUtils} from "../../mithrilUtils";
 import {MockProps} from "../../mockProps";
 import {TestModule} from "../../testModule";
