+++ src/tests/clipperUI/ratingsHelper_tests.tsx
@@ -13,7 +13,6 @@ import {ObjectUtils} from "../../scripts/objectUtils";
 import {Settings} from "../../scripts/settings";
 import {StringUtils} from "../../scripts/stringUtils";
 
-import {HelperFunctions} from "../helperFunctions";
 import {MockProps} from "../mockProps";
 import {TestModule} from "../testModule";
